requirements.txt
alembic==1.13.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dominate==2.9.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gitignore_parser==0.1.11
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
json5==0.9.25
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
postmarker==1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-frontmatter==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.36
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
visitor==0.1.3
Werkzeug==3.0.4
WTForms==3.1.2